% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calendar_dates}
\alias{calendar_dates}
\title{Create a calendar dates by specified time intervals.}
\usage{
calendar_dates(y_start, y_end, by = "1 month")
}
\arguments{
\item{y_start}{A number with the first year for the required dates.}

\item{y_end}{A number with the last year for the required dates.}

\item{by}{A number or string (as in lubridate package) to be used as interval to get the calendar dates. Default to "1 month" which is returning monthly dates (first day of the month).}
}
\value{
A vector with dates from the y_start to the y_end years at the time intervals specified by "by" argument.
}
\description{
Create a calendar dates by specified time intervals.
}
\examples{
# TBW
}
