% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.read_dsclim}
\alias{.read_dsclim}
\title{Load a cropped version of downscaled TraCE21ka .nc files}
\usage{
.read_dsclim(file, sf = NULL, proxy)
}
\arguments{
\item{file}{A string with full name (path and file name) of the .nc file to be loaded from the hard disk drive.}

\item{sf}{A sf object from the sf package to be used as template or mask to crop the downscaled trace .nc file.}

\item{proxy}{A logical value indicating whether the data should be loaded as a stars_proxy object (TRUE) or as a stars object (FALSE).}
}
\value{
A stars object with the downscaled trace data.
}
\description{
Load a cropped version of downscaled TraCE21ka .nc files
}
\examples{
# TBW
}
