% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flux2mm}
\alias{flux2mm}
\title{Flux 2 millimetters}
\usage{
flux2mm(data)
}
\arguments{
\item{data}{A stars object with precipitation in flux}
}
\value{
A stars object with precipitation in millimeters
}
\description{
Transform stars objects from precipitation in flux to precipitation in millimeters
}
\examples{
\dontrun{
data <- read_trace("data/TraCE21ka/", "PRECC")
flux2mm(data)
}
}
