% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kelvin2celsius}
\alias{kelvin2celsius}
\title{Degrees Kelvin to Celsius}
\usage{
kelvin2celsius(data)
}
\arguments{
\item{data}{A stars object with temperature in degrees Kelvin.}
}
\value{
A stars object with the temperature in degrees Celsius
}
\description{
Transform stars objects from degrees Kelvin to degrees Celsius
}
\examples{
\dontrun{
data <- read_trace("data/TraCE21ka/", "TSMX")
kelvin2celsius(data)
}
}
