% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_dsclim}
\alias{read_dsclim}
\title{Load downscaled TraCE21ka files by years.}
\usage{
read_dsclim(
  folder,
  var,
  y_start,
  y_end,
  rcp = NULL,
  gcm = NULL,
  calendar_dates = FALSE,
  sf = NULL,
  proxy = TRUE,
  ...
)
}
\arguments{
\item{folder}{A character string with the path to the folder where the downscaled TraCE21ka files, ordered in folders by variables.}

\item{var}{A character string with the name of the variables to be loaded.}

\item{y_start}{A number with the first year to be loaded.}

\item{y_end}{A number with the last year to be loaded.}

\item{rcp}{TBW}

\item{gcm}{TBW}

\item{calendar_dates}{A logical value indicating whether dates should be corrected to calendar dates.}

\item{sf}{A sf object to be used for cropping the object.}

\item{proxy}{A logical value indicating whether the data should be loaded as a stars_proxy object (TRUE) or as a stars object (FALSE).}

\item{...}{TBW}
}
\value{
A stars object with the downscaled TraCE21ka data, cropped using the sf object if provided.
}
\description{
Load downscaled TraCE21ka files by years.
}
\details{
Years should be in the format calibrated Before Present (e.g. 0 for loading year 1950 in the gregorian calendar).
}
\examples{
# TBW
}
