% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_trace}
\alias{read_trace}
\title{Read original TraCE21ka datasets}
\usage{
read_trace(folder, var, sf = NULL)
}
\arguments{
\item{folder}{TBW}

\item{var}{TBW}

\item{sf}{TBW}
}
\value{
TBW
}
\description{
Read original TraCE21ka datasets
}
\examples{
# TBW
}
