% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{time_2_calendar_dates}
\alias{time_2_calendar_dates}
\title{Modify time to calendar years}
\usage{
time_2_calendar_dates(data, y_start, y_end, by = "1 month")
}
\arguments{
\item{data}{A stars object with the data to be modified.}

\item{y_start}{A number with the starting year of the data (in calibrated Before Present format).}

\item{y_end}{A number with the ending year of the data (in calibrated Before Present format).}

\item{by}{A number or string specifying the interval of the dates to be used in the new stars object.}
}
\value{
A stars object as in data argument but with changed time dimension.
}
\description{
Modify a stars object to change the time dimension to calendar years between specified starting and ending years.
}
\examples{
# TBW
}
