% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kelvin2celsius}
\alias{kelvin2celsius}
\title{Transform stars objects from degrees kelvin to degrees celsius}
\usage{
kelvin2celsius(data)
}
\arguments{
\item{data}{A stars object with temperature as degrees kelvin.}
}
\value{
TBW
}
\description{
Transform stars objects from degrees kelvin to degrees celsius
}
\examples{
# TBW
}
