% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_trace}
\alias{read_trace}
\title{Read TraCE21ka datasets}
\usage{
read_trace(folder, var, sf = NULL)
}
\arguments{
\item{folder}{character string indicating the path to the folder with the TraCE21ka dataset.}

\item{var}{character string indicating the name of the variable that is to be readed.}

\item{sf}{a sf vector object for which the TraCE21ka data should be limited (cropped).}
}
\value{
The function returns an stars object, usually in the form of raster data, except if the argument sf is provided, in which case the object is a vector stars object.
}
\description{
Read original TraCE21ka data, returning a stars object. This function differs from the ones in the dsclim package, which return the data in the format of the climate4r framework.
}
\examples{
\dontrun{
read_trace("data/TraCE21ka/", "TSMX")
}
}
